+++ app/src/main/java/help/smartbusiness/smartaccounting/activities/CreateCreditActivity.java
@@ -10,9 +10,9 @@ import com.rengwuxian.materialedittext.MaterialAutoCompleteTextView;
 import com.rengwuxian.materialedittext.MaterialEditText;
 
 import help.smartbusiness.smartaccounting.R;
-import help.smartbusiness.smartaccounting.Utils.CustomerNameSuggester;
-import help.smartbusiness.smartaccounting.Utils.MaterialIndianCurrencyEditText;
-import help.smartbusiness.smartaccounting.Utils.Utils;
+import help.smartbusiness.smartaccounting.utils.CustomerNameSuggester;
+import help.smartbusiness.smartaccounting.utils.MaterialIndianCurrencyEditText;
+import help.smartbusiness.smartaccounting.utils.Utils;
 import help.smartbusiness.smartaccounting.models.Credit;
 import help.smartbusiness.smartaccounting.models.Customer;
