+++ app/src/main/java/help/smartbusiness/smartaccounting/services/ImportDbService.java
@@ -8,11 +8,11 @@ import android.util.Log;
 import java.io.File;
 
 import help.smartbusiness.smartaccounting.R;
-import help.smartbusiness.smartaccounting.Utils.AuthHelper;
-import help.smartbusiness.smartaccounting.Utils.DriverServicesHelper;
-import help.smartbusiness.smartaccounting.Utils.FileUtils;
-import help.smartbusiness.smartaccounting.Utils.GoogleHelper;
-import help.smartbusiness.smartaccounting.Utils.NotificationHelper;
+import help.smartbusiness.smartaccounting.utils.AuthHelper;
+import help.smartbusiness.smartaccounting.utils.DriverServicesHelper;
+import help.smartbusiness.smartaccounting.utils.FileUtils;
+import help.smartbusiness.smartaccounting.utils.GoogleHelper;
+import help.smartbusiness.smartaccounting.utils.NotificationHelper;
 import help.smartbusiness.smartaccounting.activities.BackupActivity;
 import help.smartbusiness.smartaccounting.activities.MainActivity;
 import help.smartbusiness.smartaccounting.backup.DbOperation;
