+++ app/src/main/res/layout/activity_create_credit.xml
@@ -38,7 +38,7 @@
             android:inputType="text"
             app:met_floatingLabel="highlight" />
 
-        <help.smartbusiness.smartaccounting.Utils.MaterialIndianCurrencyEditText
+        <help.smartbusiness.smartaccounting.utils.MaterialIndianCurrencyEditText
             android:id="@+id/create_credit_customer_total_due"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
