+++ app/src/main/res/layout/transaction_item_purchase_item_layout.xml
@@ -22,7 +22,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content">
 
-            <help.smartbusiness.smartaccounting.Utils.DecimalFormatterTextView
+            <help.smartbusiness.smartaccounting.utils.DecimalFormatterTextView
                 android:id="@+id/transaction_item_pi_quantity"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -30,7 +30,7 @@
                 android:textAppearance="@style/TextAppearance.AppCompat.Medium"
                 android:textColor="#FFF" />
 
-            <help.smartbusiness.smartaccounting.Utils.DecimalFormatterTextView
+            <help.smartbusiness.smartaccounting.utils.DecimalFormatterTextView
                 android:id="@+id/transaction_item_pi_rate"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -38,7 +38,7 @@
                 android:textAppearance="@style/TextAppearance.AppCompat.Medium"
                 android:textColor="#FFF" />
 
-            <help.smartbusiness.smartaccounting.Utils.CurrencyTextView
+            <help.smartbusiness.smartaccounting.utils.CurrencyTextView
                 android:id="@+id/transaction_item_pi_amount"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
